Separate _f32 and _f64 versions of dodeca functions

+++ client/src/graphics/draw.rs
@@ -438,7 +438,7 @@ impl Draw {
             for (node, transform) in nearby_nodes(
                 &sim.graph,
                 &view,
-                f64::from(self.cfg.local_simulation.view_distance),
+                self.cfg.local_simulation.view_distance,
             ) {
                 for &entity in sim.graph_entities.get(node) {
                     if sim.local_character == Some(entity) {
